fix userId Integer to Long

+++ src/main/java/com/java/team/shippingservice/repository/TokenUserRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface TokenUserRepository extends JpaRepository<TokenUser, Long> {
-    Optional<TokenUser> findByUserId(Integer userId);
+    Optional<TokenUser> findByUserId(Long userId);
 }
